Add --range option for str_replace command of edit_anthropic tool
Usually models view only a portion of the source file using --view_range option and tries to edit the file. When the file is bigger, the string-matching may not be unique and it becomes responsibility of the model to find a unique combination combining multiple lines before and after. This new option provides a way for the model to restrict the search and replacement to the provided range.

As --range option is being treated common for both view and edit comands, it's remaned from --view_range to --range.